+++ CHANGES.md
@@ -1,3 +1,12 @@
+# 1.3.0
+
+- Use synchronous SHA-256 implementation in pure JavaScript and change
+  `addMappingIdentifiers` and `mappingSamenessIdentifier` to be synchronous!
+
+# 1.2.2
+
+- Fix export of mapping functions
+
 # 1.2.1
 
 - Fix guessObjectType for registries